backend/requirements.txt
absl-py==0.8.1
aniso8601==8.0.0
astor==0.8.0
attrs==19.3.0
beautifulsoup4==4.8.1
camera==1.3.0
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
click==6.7
cryptography==2.8
cycler==0.10.0
easygui==0.98.1
eventlet==0.23.0
Flask==1.0.2
Flask-Cors==3.0.8
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SocketIO==3.0.1
Flask-WTF==0.14.2
gast==0.2.2
gevent==1.4.0
gevent-websocket==0.10.1
google-pasta==0.1.7
greenlet==0.4.15
grpcio==1.24.1
gTTS==2.0.4
gTTS-token==1.1.3
h5py==2.10.0
httplib2==0.14.0
idna==2.8
importlib-metadata==0.23
imutils==0.5.3
itsdangerous==0.24
Jinja2==2.10
joblib==0.14.0
jsonschema==3.1.1
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.1.1
matrix-client==0.3.2
more-itertools==7.2.0
numpy==1.17.3
opencv-contrib-python==4.1.1.26
opencv-python==4.1.1.26
opt-einsum==3.1.0
pdf2image==1.9.0
Pillow==6.2.0
protobuf==3.10.0
PyAudio==0.2.11
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.4.2
PyPDF2==1.26.0
pyrsistent==0.15.4
python-dateutil==2.8.0
python-engineio==2.1.1
python-socketio==1.9.0
pyttsx3==2.71
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.11.0
sklearn==0.0
soupsieve==1.9.4
SpeechRecognition==3.8.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
typing==3.7.4.1
urllib3==1.25.6
Werkzeug==0.14.1
wrapt==1.11.2
WTForms==2.2.1
zipp==0.6.0
zulip==0.6.1
